// src/utils/manifest.ts
import { Extractor } from '../extractor';
import { Source } from '../source';
import { Config, CountryCode, CustomManifest } from '../types';
import {
  disableExtractorConfigKey,
  excludeResolutionConfigKey,
  isExtractorDisabled,
  isResolutionExcluded,
} from './config';
import { envGetAppId, envGetAppName, envGetAppVersion } from './env';
import { flagFromCountryCode, languageFromCountryCode } from './language';
import { RESOLUTIONS } from './resolution';

const typedEntries = <T extends object>(obj: T): [keyof T, T[keyof T]][] => (Object.entries(obj) as [keyof T, T[keyof T]][]);

export const buildManifest = (sources: Source[], extractors: Extractor[], config: Config): CustomManifest => {
  const manifest: CustomManifest = {
    id: envGetAppId(),
    version: envGetAppVersion(),
    name: envGetAppName(),
    description: 'Provides HTTP URLs from streaming websites. Configure add-on for additional languages. Add MediaFlow proxy for more URLs.',
    resources: [
      'stream',
    ],
    types: [
      'movie',
      'series',
    ],
    catalogs: [],
    idPrefixes: ['tmdb:', 'tt'],
    logo: 'https://emojiapi.dev/api/v1/spider_web/256.png',
    behaviorHints: {
      p2p: false,
      configurable: true,
      configurationRequired: false,
    },
    config: [],
    stremioAddonsConfig: {
      issuer: 'https://stremio-addons.net',
      signature: 'eyJhbGciOiJkaXIiLCJlbmMiOiJBMTI4Q0JDLUhTMjU2In0..h1oW2E0XXKLldUqO-ReSUA.fejuyGAvmc_CdT9dnq2srZCgoC42ak-Rqeo7IKsEN3DPRpz8x-hmvbuBI_7BUU2PsFMSni35m_Lv0teUNQDPvlrm7t1FCZINMR4ty_Hee6If5m6J4kSzafD75HhWvxFU.FAcDZ5qZrTPDeRAVOUI2tQ',
    },
  };

  sources.sort((sourceA, sourceB) => sourceA.label.localeCompare(sourceB.label));

  const countryCodeSources: Partial<Record<CountryCode, Source[]>> = {};
  sources.forEach(source =>
    source.countryCodes
      .forEach(countryCode => countryCodeSources[countryCode] = [...(countryCodeSources[countryCode] ?? []), source]));

  const sortedLanguageSources = typedEntries(countryCodeSources)
    .filter(([countryCode]) => countryCode === CountryCode.pt)
    .sort(([countryCodeA], [countryCodeB]) => {
      if (countryCodeB === CountryCode.multi) {
        return 1;
      }

      return countryCodeA.localeCompare(countryCodeB);
    });

  const languages: string[] = [];
  for (const [countryCode, sources] of sortedLanguageSources) {
    const language = languageFromCountryCode(countryCode);
    languages.push(language);

    manifest.config.push({
      key: countryCode,
      type: 'checkbox',
      title: `${language} ${flagFromCountryCode(countryCode)} (${(sources as Source[]).map(source => source.label).sort().join(', ')})`,
      default: 'checked',
    });
  }

  manifest.config.push({
    key: 'showErrors',
    type: 'checkbox',
    title: 'Show errors',
    ...('showErrors' in config && { default: 'checked' }),
  });

  manifest.config.push({
    key: 'includeExternalUrls',
    type: 'checkbox',
    title: 'Include external URLs in results',
    ...('includeExternalUrls' in config && { default: 'checked' }),
  });

  manifest.config.push({
    key: 'mediaFlowProxyUrl',
    type: 'text',
    title: 'MediaFlow Proxy URL',
    default: config['mediaFlowProxyUrl'] ?? '',
  });

  manifest.config.push({
    key: 'mediaFlowProxyPassword',
    type: 'password',
    title: 'MediaFlow Proxy Password',
    default: config['mediaFlowProxyPassword'] ?? '',
  });

  RESOLUTIONS.forEach((resolution) => {
    manifest.config.push({
      key: excludeResolutionConfigKey(resolution),
      type: 'checkbox',
      title: `Exclude resolution ${resolution}`,
      ...(isResolutionExcluded(config, resolution) && { default: 'checked' }),
    });
  });

  extractors.forEach((extractor) => {
    if (extractor.id === 'external') {
      return;
    }

    manifest.config.push({
      key: disableExtractorConfigKey(extractor),
      type: 'checkbox',
      title: `Disable extractor ${extractor.label}`,
      ...(isExtractorDisabled(config, extractor) && { default: 'checked' }),
    });
  });

  manifest.description += `\n\nSupported languages: ${languages.filter(language => language !== 'Multi').join(', ')}`;
  manifest.description += `\n\nSupported sources: ${sources.map(source => source.label).join(', ')}`;
  manifest.description += `\n\nSupported extractors: ${extractors.map(extractor => extractor.label).join(', ')}`;

  return manifest;
};
